refactor(skills): reorder files
