+++ src/plugins/skills/src/behaviors/build_skill_shape/spawn_ground_target.rs
@@ -1,6 +1,7 @@
 use crate::{
 	behaviors::{SkillCaster, SkillSpawner, Target},
-	traits::skill_builder::{BuildContact, BuildProjection, LifeTimeDefinition, SkillLifetime},
+	skills::lifetime::LifeTimeDefinition,
+	traits::skill_builder::{BuildContact, BuildProjection, SkillLifetime},
 };
 use behaviors::components::ground_targeted_aoe::{
 	GroundTargetedAoeContact,
