+++ src/plugins/skills/src/skills/skill_data/run_skill_behavior_data/skill_behavior_config_data/shape_data.rs
